Added OATH Toolkit
The OATH Toolkit provide components for building one-time password
authentication systems. It contains shared libraries, command line
tools and a PAM module. Supported technologies include the event-based
HOTP algorithm (RFC4226) and the time-based TOTP algorithm (RFC6238).
OATH stands for Open AuTHentication, which is the organization that
specify the algorithms. For managing secret key files, the Portable
Symmetric Key Container (PSKC) format described in RFC6030 is supported.

The components included in the package is:

    liboath: A shared and static C library for OATH handling.

    oathtool: A command line tool for generating and validating OTPs.

http://www.nongnu.org/oath-toolkit/